requirements.txt
streamlit
streamlit-geolocation
gspread
oauth2client
pandas
requests
geopy
pydeck
streamlit-autorefresh
google-autorefresh
pytz